+++ packages/integration-sdk/CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to
 
 ## Unreleased
 
+### Changed
+
+- `publishMetric` now logs the metric that is published.
+
 ## 2.1.1 - 2020-06-23
 
 ### Changed